requirements.txt
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
axios==0.4.0
bcrypt==4.1.2
billiard==4.2.0
blinker==1.7.0
cachelib==0.9.0
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cors==1.0.1
dnspython==2.4.2
email-validator==2.1.0.post1
filelock==3.13.1
Flask==3.0.1
Flask-Caching==2.1.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
future==0.18.3
gevent==23.9.1
greenlet==3.0.3
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
passlib==1.7.4
prompt-toolkit==3.0.43
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
requests-file==1.5.1
rich==13.7.0
six==1.16.0
SQLAlchemy==2.0.25
tldextract==5.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.1
WTForms==3.1.2
zope.event==5.0
zope.interface==6.1
